update docs, and move client to separate file
